Updated to use PtrWrapper around Parameters.
I'm not sure that this is strictly needed or desirable, but it
at least makes things consistent.

When written with just raw Parameters (not pointer) I was getting some
valgrind errors, and the physiological calculations were broken.  This looked
to be to do with some copying that was leaving the internal state of Parameters
confused.  Probably would have been desirable to fix that, but this approach
also works, and is as clear.