Change logic of height calculation for empty Species/Patches
Because a light environment spline will be needed to compute the
germination probability, we need a light environment that is known
at least to the height of a seed (even though this will always be 1
everywhere).  This changes the logic around height_max() so that it
will return the height of a seed/seedling for Species where there no
individuals.