Use Environment rather than splines for physiogical calculations.
This affects a lot of different parts, but perhaps the tests more
than the actual compiled code.  Plants (and their descendants) now
take the Environment class for updating physiology.  This allows
us to pass in a bit of extra information and abstract away some of the
gory details.

Because making Environment objects is a bit more work than making a
plain Spline, I've made a small testing environment generation function
within helper-tree.R.  This builds the environment I was using everywhere
already.