AdaptiveSpline interface same as other utility classes.
The constructor now takes control parameters (tolerances, etc), and
the function that generates the spline takes the functor.  This makes
the function behave like FindRoot and Integrator.

This fixes #18

At the same time, the Environment can now initialise the AdaptiveSpline
generator with the appropriate control parameters, and then build
light environment splines on request.  This means that the lifetime of the
AdaptiveSpline is the same as Environment, and will save a nontrivial
amount of time with construction/destruction.