Disturbance handling simplified.
Rather than keep the 'Disturbance' object in Parameters, we now
just keep its parameters.  When Environment is built, it will
construct a Disturbance object, rather than copying the one that
the Parameters had.

This makes it much easier to modify things at the R level, and
seems a lot more sensible.

Fixes #34