Simplify seed_rain handling.
Now stored as a vector within the Parameters, rather than set
specially in the controlling classes.  This further enforces the
idea that things should not change during a simulation, without
requiring all the checks from before.

This change is really needed before the comparisons against the
reference implementation can be done.

Fixes #33