EBT's use of CohortSchedule now much more obvious.
Now, there are a series of events with a start and an end time.
The EBT checks that we are actually at the expected start time,
adds a cohort, then advances to the end time.  This removes
the need for the special (final) event, which will be removed
in the next commit.

One possible issue with this approach (but also any approach) is
that no time passes when a cohort is added.  So the "state" of
the system might be the state right before a new cohort is
added, or right after.  In the previous version, it was the state
right after.  Now, it's the state right before.