Two major modifications of "density" handling for CohortTop
First, log_density is now *actually* log_density (previously it
was negative log_density, though a few versions ago it was
called "density" even though it was log_density).  This affects
the derivative calculation (change in sign) and the only place
where density is used (change in sign on use in
CohortTop::leaf_area_above).

The other is a bug in that the germination probability should have
been multiplied into the incoming seed rain (in
CohortTop::compute_initial_conditions).  Documentation for this
change will be added to doc/EBT.md shortly (Daniel has uncommited
changes).