Support for using runsteady to compute equilibrium seed rain
This is still slow, taking many more iterations than would be nice.  But
it does seem the most robust I've seen.  A better approach might be to
use this to get close to the equilibrium and then use a rootsolving
approach to polish the root, but I don't know how much better that would
be in practice.