Big rename of model components
Bits that are specific to a physiology are now prefixed with FFW16_
(includes Strategy, Parameters, Plant, Cohort, Species, Patch, EBT).
Some, but not all, of these now have proper template based generation
on the R side, but I need to fix support for list classes in RcppR6
before finishing this off.