Move the heartwood variables into Plant
This narrows the gap between Plant and PlantPlus and will make
modularisation much easier.  This changes a *lot* of tests
(primarily because of the change in ODE dimension), so will
probably affect downstream things a lot.

It should now also be trivial to construct a PlantPlus from a
Plant (because the variables are very similar), so this will
reduce how important PlantPlus objects will be.