All classes templated
This is not tested, but this means that adding new classes to run
with the EBT *should* be as simple as adding a line into the
concrete definitions for:

* Plant
* Parameters
* Cohort
* Patch
* EBT
* StochasticSpecies
* StochasticPatch
* StochasticPatchRunner

(the last three being optional).  _all_ the others must be provided,
as Patches return Cohorts return Plants, so we need to be able to
compile all C++ -> R conversions.