Fix recycling bug
The narea calculations assumed that there was only a single strategy, but
the hyperparametersiation function is vectorised.  The warning was pretty
harmless in the cases where narea was a scalar but would have led to
nonsensical results where narea is a vector.  It would be worth adding some
tests (e.g. into test-parameters.R) to check that the narea calculations
are actually working as expected (@dfalster)