Deal better with integration failure in grow_plant_to_size
During an integration failure, assume that plants have failed to grow,
most likely due to impossibly large mortality.  Treat this the same way
as failing to reach a given size and discard all larger sizes.  Includes
a check to make sure that it is mortality that is the problem but this
is not really tested yet.

Fixes #174