Updating to use pkgdown for website construction.
# For issue 185
We previously were using the staticdocs package to build the website but that package is no longer supported.
pkgdown builds into the docs folder so content in the old docs folder can now be found at inst/docs.

To build the site we run make pkgdown instead of make staticdocs.