Scaffolder uses any template, documentation updates, test-reference-strateg-ff16.R
As per #183 the scaffolder can now  take any strategy and then copy this strategy with a new name, this will speed up the creation of new strategies. Documentation in the developer notes was also updated.
Some tests were moved from test-strategy-ff16.R to test-reference-comparison.R to enable the scaffolder to use ff16 as a template.
Documentation in scm_support.R was also updated to be more generic.