Update plantRunner tests to work with FF16r strategy
With the extensions to PlantRunner to handle multiple strategies,
the tests for fFF16r strategy were failing. I discovered that
fecundity in this strategy at the specified heights was 0, which is
allowable. So tests have been updated to allow for the fact that
this variable can be also equal to the comparison value.