Add missing tests for fitness & equilibrium functions (#355)
* Missing tests on fitness
* Rename test file so comes later in test routines

Despite extensive tests, somehow various functions related to fitness and demographic equilibrium went untested. Presumably because they were slow. Lack of tests here enabled loss of functionality as we changed some of the interface.

This PR adds some tests on master, which will then be ported back to develop and updated as needed.